--- zebra-chain/src/orchard_zsa/common.rs
@@ -1,23 +1,0 @@
-//! Serialization implementation for selected types from the 'orchard_zsa' crate used in this module.
-
-use std::io;
-
-use crate::serialization::{ReadZcashExt, SerializationError, ZcashDeserialize, ZcashSerialize};
-
-use orchard::note::AssetBase;
-
-// The size of the serialized AssetBase in bytes (used for TrustedPreallocate impls)
-pub(super) const ASSET_BASE_SIZE: u64 = 32;
-
-impl ZcashSerialize for AssetBase {
-    fn zcash_serialize<W: io::Write>(&self, mut writer: W) -> Result<(), io::Error> {
-        writer.write_all(&self.to_bytes())
-    }
-}
-
-impl ZcashDeserialize for AssetBase {
-    fn zcash_deserialize<R: io::Read>(mut reader: R) -> Result<Self, SerializationError> {
-        Option::from(AssetBase::from_bytes(&reader.read_32_bytes()?))
-            .ok_or_else(|| SerializationError::Parse("Invalid orchard_zsa AssetBase!"))
-    }
-}
